Move generate_signed_url to standalone function in credentials.
This was **very** out of place as a method on Connection in storage.
