+++ gcloud/storage/test_connection.py
@@ -602,43 +602,6 @@ class TestConnection(unittest2.TestCase):
         conn = self._makeOne(PROJECT)
         self.assertRaises(TypeError, conn.new_bucket, object())
 
-    def test_generate_signed_url_w_expiration_int(self):
-        import base64
-        import urlparse
-        from gcloud._testing import _Monkey
-        from gcloud.test_credentials import _Credentials
-        from gcloud.storage import connection as MUT
-
-        ENDPOINT = 'http://api.example.com'
-        RESOURCE = '/name/path'
-        PROJECT = 'project'
-        SIGNED = base64.b64encode('DEADBEEF')
-        conn = self._makeOne(PROJECT, _Credentials())
-        conn.API_ACCESS_ENDPOINT = ENDPOINT
-
-        def _get_signed_query_params(*args):
-            credentials, expiration = args[:2]
-            return {
-                'GoogleAccessId': credentials.service_account_name,
-                'Expires': str(expiration),
-                'Signature': SIGNED,
-            }
-
-        with _Monkey(MUT, _get_signed_query_params=_get_signed_query_params):
-            url = conn.generate_signed_url(RESOURCE, 1000)
-
-        scheme, netloc, path, qs, frag = urlparse.urlsplit(url)
-        self.assertEqual(scheme, 'http')
-        self.assertEqual(netloc, 'api.example.com')
-        self.assertEqual(path, RESOURCE)
-        params = urlparse.parse_qs(qs)
-        self.assertEqual(len(params), 3)
-        self.assertEqual(params['Signature'], [SIGNED])
-        self.assertEqual(params['Expires'], ['1000'])
-        self.assertEqual(params['GoogleAccessId'],
-                         [_Credentials.service_account_name])
-        self.assertEqual(frag, '')
-
 
 class Test__BucketIterator(unittest2.TestCase):
 
@@ -676,225 +639,6 @@ class Test__BucketIterator(unittest2.TestCase):
         self.assertEqual(bucket.name, BLOB_NAME)
 
 
-class Test__get_expiration_seconds(unittest2.TestCase):
-
-    def _callFUT(self, expiration):
-        from gcloud.storage.connection import _get_expiration_seconds
-
-        return _get_expiration_seconds(expiration)
-
-    def _utc_seconds(self, when):
-        import calendar
-
-        return int(calendar.timegm(when.timetuple()))
-
-    def test__get_expiration_seconds_w_invalid(self):
-        self.assertRaises(TypeError, self._callFUT, object())
-        self.assertRaises(TypeError, self._callFUT, None)
-
-    def test__get_expiration_seconds_w_int(self):
-        self.assertEqual(self._callFUT(123), 123)
-
-    def test__get_expiration_seconds_w_long(self):
-        try:
-            long
-        except NameError:  # pragma: NO COVER Py3K
-            pass
-        else:
-            self.assertEqual(self._callFUT(long(123)), 123)
-
-    def test__get_expiration_w_naive_datetime(self):
-        import datetime
-
-        expiration_no_tz = datetime.datetime(2004, 8, 19, 0, 0, 0, 0)
-        utc_seconds = self._utc_seconds(expiration_no_tz)
-        self.assertEqual(self._callFUT(expiration_no_tz), utc_seconds)
-
-    def test__get_expiration_w_utc_datetime(self):
-        import datetime
-        import pytz
-
-        expiration_utc = datetime.datetime(2004, 8, 19, 0, 0, 0, 0, pytz.utc)
-        utc_seconds = self._utc_seconds(expiration_utc)
-        self.assertEqual(self._callFUT(expiration_utc), utc_seconds)
-
-    def test__get_expiration_w_other_zone_datetime(self):
-        import datetime
-        import pytz
-
-        zone = pytz.timezone('CET')
-        expiration_other = datetime.datetime(2004, 8, 19, 0, 0, 0, 0, zone)
-        utc_seconds = self._utc_seconds(expiration_other)
-        cet_seconds = utc_seconds - (60 * 60)  # CET one hour earlier than UTC
-        self.assertEqual(self._callFUT(expiration_other), cet_seconds)
-
-    def test__get_expiration_seconds_w_timedelta_seconds(self):
-        import datetime
-        from gcloud.storage import connection
-        from gcloud._testing import _Monkey
-
-        dummy_utcnow = datetime.datetime(2004, 8, 19, 0, 0, 0, 0)
-        utc_seconds = self._utc_seconds(dummy_utcnow)
-        expiration_as_delta = datetime.timedelta(seconds=10)
-
-        with _Monkey(connection, _utcnow=lambda: dummy_utcnow):
-            result = self._callFUT(expiration_as_delta)
-
-        self.assertEqual(result, utc_seconds + 10)
-
-    def test__get_expiration_seconds_w_timedelta_days(self):
-        import datetime
-        from gcloud.storage import connection
-        from gcloud._testing import _Monkey
-
-        dummy_utcnow = datetime.datetime(2004, 8, 19, 0, 0, 0, 0)
-        utc_seconds = self._utc_seconds(dummy_utcnow)
-        expiration_as_delta = datetime.timedelta(days=1)
-
-        with _Monkey(connection, _utcnow=lambda: dummy_utcnow):
-            result = self._callFUT(expiration_as_delta)
-
-        self.assertEqual(result, utc_seconds + 86400)
-
-
-class Test__get_pem_key(unittest2.TestCase):
-
-    def _callFUT(self, credentials):
-        from gcloud.storage.connection import _get_pem_key
-        return _get_pem_key(credentials)
-
-    def test_bad_argument(self):
-        self.assertRaises(TypeError, self._callFUT, None)
-
-    def test_signed_jwt_for_p12(self):
-        import base64
-        from oauth2client import client
-        from gcloud._testing import _Monkey
-        from gcloud.storage import connection as MUT
-
-        scopes = []
-        PRIVATE_KEY = 'dummy_private_key_text'
-        credentials = client.SignedJwtAssertionCredentials(
-            'dummy_service_account_name', PRIVATE_KEY, scopes)
-        crypt = _Crypt()
-        rsa = _RSA()
-        with _Monkey(MUT, crypt=crypt, RSA=rsa):
-            result = self._callFUT(credentials)
-
-        self.assertEqual(crypt._private_key_text,
-                         base64.b64encode(PRIVATE_KEY))
-        self.assertEqual(crypt._private_key_password, 'notasecret')
-        self.assertEqual(result, 'imported:__PEM__')
-
-    def test_service_account_via_json_key(self):
-        from oauth2client import service_account
-        from gcloud._testing import _Monkey
-        from gcloud.storage import connection as MUT
-
-        scopes = []
-
-        PRIVATE_TEXT = 'dummy_private_key_pkcs8_text'
-
-        def _get_private_key(private_key_pkcs8_text):
-            return private_key_pkcs8_text
-
-        with _Monkey(service_account, _get_private_key=_get_private_key):
-            credentials = service_account._ServiceAccountCredentials(
-                'dummy_service_account_id', 'dummy_service_account_email',
-                'dummy_private_key_id', PRIVATE_TEXT, scopes)
-
-        rsa = _RSA()
-        with _Monkey(MUT, RSA=rsa):
-            result = self._callFUT(credentials)
-
-        expected = 'imported:%s' % (PRIVATE_TEXT,)
-        self.assertEqual(result, expected)
-
-
-class Test__get_signed_query_params(unittest2.TestCase):
-
-    def _callFUT(self, credentials, expiration, signature_string):
-        from gcloud.storage.connection import _get_signed_query_params
-        return _get_signed_query_params(credentials, expiration,
-                                        signature_string)
-
-    def test_wrong_type(self):
-        from gcloud._testing import _Monkey
-        from gcloud.storage import connection as MUT
-
-        pkcs_v1_5 = _PKCS1_v1_5()
-        rsa = _RSA()
-        sha256 = _SHA256()
-
-        def _get_pem_key(credentials):
-            return credentials
-
-        BAD_CREDENTIALS = None
-        EXPIRATION = '100'
-        SIGNATURE_STRING = 'dummy_signature'
-        with _Monkey(MUT, RSA=rsa, PKCS1_v1_5=pkcs_v1_5,
-                     SHA256=sha256, _get_pem_key=_get_pem_key):
-            self.assertRaises(NameError, self._callFUT,
-                              BAD_CREDENTIALS, EXPIRATION, SIGNATURE_STRING)
-
-    def _run_test_with_credentials(self, credentials, account_name):
-        import base64
-        from gcloud._testing import _Monkey
-        from gcloud.storage import connection as MUT
-
-        crypt = _Crypt()
-        pkcs_v1_5 = _PKCS1_v1_5()
-        rsa = _RSA()
-        sha256 = _SHA256()
-
-        EXPIRATION = '100'
-        SIGNATURE_STRING = 'dummy_signature'
-        with _Monkey(MUT, crypt=crypt, RSA=rsa, PKCS1_v1_5=pkcs_v1_5,
-                     SHA256=sha256):
-            result = self._callFUT(credentials, EXPIRATION, SIGNATURE_STRING)
-
-        if crypt._pkcs12_key_as_pem_called:
-            self.assertEqual(crypt._private_key_text,
-                             base64.b64encode('dummy_private_key_text'))
-            self.assertEqual(crypt._private_key_password, 'notasecret')
-        self.assertEqual(sha256._signature_string, SIGNATURE_STRING)
-        SIGNED = base64.b64encode('DEADBEEF')
-        expected_query = {
-            'Expires': EXPIRATION,
-            'GoogleAccessId': account_name,
-            'Signature': SIGNED,
-        }
-        self.assertEqual(result, expected_query)
-
-    def test_signed_jwt_for_p12(self):
-        from oauth2client import client
-
-        scopes = []
-        ACCOUNT_NAME = 'dummy_service_account_name'
-        credentials = client.SignedJwtAssertionCredentials(
-            ACCOUNT_NAME, 'dummy_private_key_text', scopes)
-        self._run_test_with_credentials(credentials, ACCOUNT_NAME)
-
-    def test_service_account_via_json_key(self):
-        from oauth2client import service_account
-        from gcloud._testing import _Monkey
-
-        scopes = []
-
-        PRIVATE_TEXT = 'dummy_private_key_pkcs8_text'
-
-        def _get_private_key(private_key_pkcs8_text):
-            return private_key_pkcs8_text
-
-        ACCOUNT_NAME = 'dummy_service_account_email'
-        with _Monkey(service_account, _get_private_key=_get_private_key):
-            credentials = service_account._ServiceAccountCredentials(
-                'dummy_service_account_id', ACCOUNT_NAME,
-                'dummy_private_key_id', PRIVATE_TEXT, scopes)
-
-        self._run_test_with_credentials(credentials, ACCOUNT_NAME)
-
-
 class Http(object):
 
     _called_with = None
@@ -907,45 +651,3 @@ class Http(object):
     def request(self, **kw):
         self._called_with = kw
         return self._response, self._content
-
-
-class _Crypt(object):
-
-    _pkcs12_key_as_pem_called = False
-
-    def pkcs12_key_as_pem(self, private_key_text, private_key_password):
-        self._pkcs12_key_as_pem_called = True
-        self._private_key_text = private_key_text
-        self._private_key_password = private_key_password
-        return '__PEM__'
-
-
-class _RSA(object):
-
-    _imported = None
-
-    def importKey(self, pem):
-        self._imported = pem
-        return 'imported:%s' % pem
-
-
-class _PKCS1_v1_5(object):
-
-    _pem_key = _signature_hash = None
-
-    def new(self, pem_key):
-        self._pem_key = pem_key
-        return self
-
-    def sign(self, signature_hash):
-        self._signature_hash = signature_hash
-        return 'DEADBEEF'
-
-
-class _SHA256(object):
-
-    _signature_string = None
-
-    def new(self, signature_string):
-        self._signature_string = signature_string
-        return self
